refactor: extract types-generator to separate file
