+++ packages/astro/src/content/vite-plugin-content-server.ts
@@ -1,304 +1,46 @@
-import { Plugin, normalizePath } from 'vite';
-import glob from 'fast-glob';
+import type { Plugin } from 'vite';
 import fsMod from 'node:fs';
-import * as path from 'node:path';
-import { bold, cyan } from 'kleur/colors';
-import { info, LogOptions, warn } from '../core/logger/core.js';
+import { cyan } from 'kleur/colors';
+import { info, LogOptions } from '../core/logger/core.js';
 import type { AstroSettings } from '../@types/astro.js';
-import { appendForwardSlash, isRelativePath } from '../core/path.js';
-import { contentFileExts, CONTENT_FLAG, CONTENT_TYPES_FILE } from './consts.js';
+import { contentFileExts, CONTENT_FLAG } from './consts.js';
 import { escapeViteEnvReferences } from '../vite-plugin-utils/index.js';
-import { pathToFileURL } from 'node:url';
 import {
-	CollectionConfig,
-	ContentConfig,
 	getEntryData,
 	getEntrySlug,
 	getContentPaths,
 	loadContentConfig,
-	NotFoundError,
-	observable,
+	contentObservable,
 	parseFrontmatter,
 	ContentPaths,
 } from './utils.js';
 import * as devalue from 'devalue';
-
-type ChokidarEvent = 'add' | 'addDir' | 'change' | 'unlink' | 'unlinkDir';
-type ContentEvent = { name: ChokidarEvent; entry: string };
-type EntryInfo = {
-	id: string;
-	slug: string;
-	collection: string;
-};
-
-type GenerateContent = {
-	init(): Promise<void>;
-	queueEvent(event: ContentEvent): void;
-};
-
-type ContentTypes = Record<string, Record<string, string>>;
-
-const msg = {
-	collectionAdded: (collection: string) => `${cyan(collection)} collection added`,
-	entryAdded: (entry: string, collection: string) => `${cyan(entry)} added to ${bold(collection)}.`,
-};
+import {
+	createContentTypesGenerator,
+	GenerateContentTypes,
+	getEntryInfo,
+	getEntryType,
+} from './types-generator.js';
 
 interface AstroContentServerPluginParams {
 	fs: typeof fsMod;
 	logging: LogOptions;
 	settings: AstroSettings;
-	contentConfig: ContentConfig | Error;
 	mode: string;
-}
-
-function warnNonexistentCollections({
-	contentConfig,
-	contentTypes,
-	logging,
-}: {
-	contentConfig: ContentConfig;
-	contentTypes: ContentTypes;
-	logging: LogOptions;
-}) {
-	for (const configuredCollection in contentConfig.collections) {
-		if (!contentTypes[JSON.stringify(configuredCollection)]) {
-			warn(
-				logging,
-				'content',
-				`${JSON.stringify(
-					configuredCollection
-				)} is not a collection. Check your content config for typos.`
-			);
-		}
-	}
 }
 
 export function astroContentServerPlugin({
 	fs,
 	settings,
-	contentConfig,
 	logging,
 	mode,
 }: AstroContentServerPluginParams): Plugin[] {
 	const contentPaths: ContentPaths = getContentPaths({ srcDir: settings.config.srcDir });
 	let contentDirExists = false;
-	let contentGenerator: GenerateContent;
-	let configFileLoading = observable({ isLoading: false });
-
-	async function createContentGenerator(): Promise<GenerateContent> {
-		const contentTypes: ContentTypes = {};
-
-		let events: Promise<void>[] = [];
-		let debounceTimeout: NodeJS.Timeout | undefined;
-		let eventsSettled: Promise<void> | undefined;
-		let shouldWriteTypesFile = false;
-
-		const contentTypesBase = await fsMod.promises.readFile(
-			new URL(CONTENT_TYPES_FILE, contentPaths.generatedInputDir),
-			'utf-8'
-		);
-
-		async function init() {
-			shouldWriteTypesFile = true;
-			const pattern =
-				new URL('./**/', contentPaths.contentDir).pathname + `*{${contentFileExts.join(',')}}`;
-			const entries = await glob(pattern, {
-				fs: {
-					readdir: fs.readdir.bind(fs),
-					readdirSync: fs.readdirSync.bind(fs),
-				},
-			});
-			for (const entry of entries) {
-				queueEvent({ name: 'add', entry }, { shouldLog: false });
-			}
-			await eventsSettled;
-			if (!(contentConfig instanceof Error)) {
-				warnNonexistentCollections({ logging, contentConfig, contentTypes });
-			}
-		}
-
-		async function onEvent(event: ContentEvent, opts?: { shouldLog: boolean }) {
-			const shouldLog = opts?.shouldLog ?? true;
-
-			if (event.name === 'addDir' || event.name === 'unlinkDir') {
-				const collection = path.relative(contentPaths.contentDir.pathname, event.entry);
-				// If directory is multiple levels deep, it is not a collection!
-				const isCollectionEvent = collection.split(path.sep).length === 1;
-				if (!isCollectionEvent) return;
-				switch (event.name) {
-					case 'addDir':
-						addCollection(contentTypes, JSON.stringify(collection));
-						if (shouldLog) {
-							info(logging, 'content', msg.collectionAdded(collection));
-						}
-						break;
-					case 'unlinkDir':
-						removeCollection(contentTypes, JSON.stringify(collection));
-						break;
-				}
-			} else {
-				const fileType = getEntryType(event.entry, contentPaths);
-				if (fileType === 'config') {
-					configFileLoading.ctx.isLoading = true;
-					contentConfig = await loadContentConfig({ fs, settings });
-					configFileLoading.ctx.isLoading = false;
-					configFileLoading.notify();
-					configFileLoading.clear();
-					if (!(contentConfig instanceof Error)) {
-						warnNonexistentCollections({ logging, contentConfig, contentTypes });
-					}
-					return;
-				}
-				if (fileType === 'unknown') {
-					warn(
-						logging,
-						'content',
-						`${cyan(
-							path.relative(contentPaths.contentDir.pathname, event.entry)
-						)} is not a supported file type. Skipping.`
-					);
-					return;
-				}
-				const entryInfo = getEntryInfo({
-					entryPath: event.entry,
-					contentDir: contentPaths.contentDir,
-				});
-				// Not a valid `src/content/` entry. Silently return, but should be impossible?
-				if (entryInfo instanceof Error) return;
-				if (entryInfo.collection === '.') {
-					warn(
-						logging,
-						'content',
-						`${cyan(
-							path.relative(contentPaths.contentDir.pathname, event.entry)
-						)} must be nested in a collection directory. Skipping.`
-					);
-					return;
-				}
-
-				const { id, slug, collection } = entryInfo;
-				const collectionKey = JSON.stringify(collection);
-				const entryKey = JSON.stringify(id);
-				const collectionConfig =
-					contentConfig instanceof Error ? undefined : contentConfig.collections[collection];
-				switch (event.name) {
-					case 'add':
-						shouldWriteTypesFile = true;
-						if (!(collectionKey in contentTypes)) {
-							addCollection(contentTypes, collectionKey);
-						}
-						if (!(entryKey in contentTypes[collectionKey])) {
-							addEntry(contentTypes, collectionKey, entryKey, slug, collectionConfig);
-						}
-						if (shouldLog) {
-							info(logging, 'content', msg.entryAdded(entryInfo.slug, entryInfo.collection));
-						}
-						break;
-					case 'unlink':
-						shouldWriteTypesFile = true;
-						if (collectionKey in contentTypes && entryKey in contentTypes[collectionKey]) {
-							removeEntry(contentTypes, collectionKey, entryKey);
-						}
-						break;
-					case 'change':
-						// noop. Frontmatter types are inferred from collection schema import, so they won't change!
-						break;
-				}
-			}
-		}
-
-		function queueEvent(event: ContentEvent, eventOpts?: { shouldLog: boolean }) {
-			if (!event.entry.startsWith(contentPaths.contentDir.pathname)) return;
-			if (event.entry.endsWith(CONTENT_TYPES_FILE)) return;
-
-			events.push(onEvent(event, eventOpts));
-			runEventsDebounced();
-		}
-
-		function runEventsDebounced() {
-			eventsSettled = new Promise((resolve, reject) => {
-				try {
-					debounceTimeout && clearTimeout(debounceTimeout);
-					debounceTimeout = setTimeout(async () => {
-						await Promise.all(events);
-						if (shouldWriteTypesFile) {
-							await writeContentFiles({
-								fs,
-								contentTypes,
-								paths: contentPaths,
-								contentTypesBase,
-								hasContentConfig: !(contentConfig instanceof NotFoundError),
-							});
-							shouldWriteTypesFile = false;
-						}
-						resolve();
-					}, 50 /* debounce 50 ms to batch chokidar events */);
-				} catch (e) {
-					reject(e);
-				}
-			});
-		}
-		return { init, queueEvent };
-	}
+	let contentGenerator: GenerateContentTypes;
+	const contentConfigObserver = contentObservable({ status: 'loading' });
 
 	return [
-		{
-			name: 'content-flag-plugin',
-			async load(id) {
-				const { pathname, searchParams } = new URL(id, 'file://');
-				if (isContentFlagImport({ pathname, searchParams })) {
-					if (configFileLoading.ctx.isLoading) {
-						await new Promise((resolve) => {
-							configFileLoading.subscribe(resolve);
-						});
-					}
-					const rawContents = await fs.promises.readFile(pathname, 'utf-8');
-					const {
-						content: body,
-						data: unparsedData,
-						matter: rawData = '',
-					} = parseFrontmatter(rawContents, pathname);
-					const entryInfo = getEntryInfo({
-						entryPath: pathname,
-						contentDir: contentPaths.contentDir,
-					});
-					if (entryInfo instanceof Error) return;
-
-					const _internal = { filePath: pathname, rawData };
-					const partialEntry = { data: unparsedData, body, _internal, ...entryInfo };
-					const collectionConfig =
-						contentConfig instanceof Error
-							? undefined
-							: contentConfig.collections[entryInfo.collection];
-					const data = collectionConfig
-						? await getEntryData(partialEntry, collectionConfig)
-						: unparsedData;
-					const slug = collectionConfig
-						? await getEntrySlug({ ...partialEntry, data }, collectionConfig)
-						: entryInfo.slug;
-
-					const code = escapeViteEnvReferences(`
-export const id = ${JSON.stringify(entryInfo.id)};
-export const collection = ${JSON.stringify(entryInfo.collection)};
-export const slug = ${JSON.stringify(slug)};
-export const body = ${JSON.stringify(body)};
-export const data = ${devalue.uneval(data) /* TODO: reuse astro props serializer */};
-export const _internal = {
-	filePath: ${JSON.stringify(pathname)},
-	rawData: ${JSON.stringify(rawData)},
-};
-`);
-					return { code };
-				}
-			},
-			async transform(code, id) {
-				if (isContentFlagImport(new URL(id, 'file://'))) {
-					// Escape before Rollup internal transform.
-					// Base on MUCH trial-and-error, inspired by MDX integration 2-step transform.
-					return { code: escapeViteEnvReferences(code) };
-				}
-			},
-		},
 		{
 			name: 'astro-content-server-plugin',
 			async config(viteConfig) {
@@ -310,9 +52,23 @@ export const _internal = {
 					return;
 				}
 
+				contentConfigObserver.set({ status: 'loading' });
+				const config = await loadContentConfig({ fs, settings });
+				if (config instanceof Error) {
+					contentConfigObserver.set({ status: 'error', error: config });
+				} else {
+					contentConfigObserver.set({ status: 'loaded', config });
+				}
+
 				if (mode === 'dev' || viteConfig.build?.ssr === true) {
 					info(logging, 'content', 'Generating entries...');
-					contentGenerator = await createContentGenerator();
+					contentGenerator = await createContentTypesGenerator({
+						fs,
+						settings,
+						logging,
+						contentConfigObserver,
+						contentPaths,
+					});
 					await contentGenerator.init();
 				}
 			},
@@ -372,115 +128,72 @@ export const _internal = {
 				}
 			},
 		},
+		{
+			name: 'astro-content-flag-plugin',
+			async load(id) {
+				const { pathname, searchParams } = new URL(id, 'file://');
+				if (isContentFlagImport({ pathname, searchParams })) {
+					if (contentConfigObserver.get().status === 'loading') {
+						await new Promise((resolve) => {
+							const unsubscribe = contentConfigObserver.subscribe((ctx) => {
+								if (ctx.status === 'loaded') {
+									unsubscribe();
+									resolve(undefined);
+								}
+							});
+						});
+					}
+					const observable = contentConfigObserver.get();
+					const rawContents = await fs.promises.readFile(pathname, 'utf-8');
+					const {
+						content: body,
+						data: unparsedData,
+						matter: rawData = '',
+					} = parseFrontmatter(rawContents, pathname);
+					const entryInfo = getEntryInfo({
+						entryPath: pathname,
+						contentDir: contentPaths.contentDir,
+					});
+					if (entryInfo instanceof Error) return;
+
+					const _internal = { filePath: pathname, rawData };
+					const partialEntry = { data: unparsedData, body, _internal, ...entryInfo };
+					const collectionConfig =
+						observable.status === 'loaded'
+							? observable.config.collections[entryInfo.collection]
+							: undefined;
+					const data = collectionConfig
+						? await getEntryData(partialEntry, collectionConfig)
+						: unparsedData;
+					const slug = collectionConfig
+						? await getEntrySlug({ ...partialEntry, data }, collectionConfig)
+						: entryInfo.slug;
+
+					const code = escapeViteEnvReferences(`
+export const id = ${JSON.stringify(entryInfo.id)};
+export const collection = ${JSON.stringify(entryInfo.collection)};
+export const slug = ${JSON.stringify(slug)};
+export const body = ${JSON.stringify(body)};
+export const data = ${devalue.uneval(data) /* TODO: reuse astro props serializer */};
+export const _internal = {
+	filePath: ${JSON.stringify(pathname)},
+	rawData: ${JSON.stringify(rawData)},
+};
+`);
+					return { code };
+				}
+			},
+			async transform(code, id) {
+				if (isContentFlagImport(new URL(id, 'file://'))) {
+					// Escape before Rollup internal transform.
+					// Base on MUCH trial-and-error, inspired by MDX integration 2-step transform.
+					return { code: escapeViteEnvReferences(code) };
+				}
+			},
+		},
 	];
 }
 
 function isContentFlagImport({ searchParams, pathname }: Pick<URL, 'searchParams' | 'pathname'>) {
 	return searchParams.has(CONTENT_FLAG) && contentFileExts.some((ext) => pathname.endsWith(ext));
-}
-
-function addCollection(contentMap: ContentTypes, collectionKey: string) {
-	contentMap[collectionKey] = {};
-}
-
-function removeCollection(contentMap: ContentTypes, collectionKey: string) {
-	delete contentMap[collectionKey];
-}
-
-function addEntry(
-	contentTypes: ContentTypes,
-	collectionKey: string,
-	entryKey: string,
-	slug: string,
-	collectionConfig?: CollectionConfig
-) {
-	const dataType = collectionConfig?.schema ? `InferEntrySchema<${collectionKey}>` : 'any';
-	// If user has custom slug function, we can't predict slugs at type compilation.
-	// Would require parsing all data and evaluating ahead-of-time;
-	// We evaluate with lazy imports at dev server runtime
-	// to prevent excessive errors
-	const slugType = collectionConfig?.slug ? 'string' : JSON.stringify(slug);
-
-	contentTypes[collectionKey][
-		entryKey
-	] = `{\n  id: ${entryKey},\n  slug: ${slugType},\n  body: string,\n  collection: ${collectionKey},\n  data: ${dataType}\n}`;
-}
-
-function removeEntry(contentTypes: ContentTypes, collectionKey: string, entryKey: string) {
-	delete contentTypes[collectionKey][entryKey];
-}
-
-function getEntryInfo({
-	entryPath,
-	contentDir,
-}: Pick<ContentPaths, 'contentDir'> & { entryPath: string }): EntryInfo | Error {
-	const relativeEntryPath = normalizePath(path.relative(contentDir.pathname, entryPath));
-	const collection = path.dirname(relativeEntryPath).split(path.sep).shift();
-	if (!collection) return new Error();
-
-	const id = path.relative(collection, relativeEntryPath);
-	const slug = id.replace(path.extname(id), '');
-	return {
-		id,
-		slug,
-		collection,
-	};
-}
-
-function getEntryType(entryPath: string, paths: ContentPaths): 'content' | 'config' | 'unknown' {
-	const { dir, ext, name } = path.parse(entryPath);
-	const { pathname } = new URL(name, appendForwardSlash(pathToFileURL(dir).href));
-	if ((contentFileExts as readonly string[]).includes(ext)) {
-		return 'content';
-	} else if (pathname === paths.config.pathname) {
-		return 'config';
-	} else {
-		return 'unknown';
-	}
-}
-
-async function writeContentFiles({
-	fs,
-	paths,
-	contentTypes,
-	contentTypesBase,
-	hasContentConfig,
-}: {
-	fs: typeof fsMod;
-	paths: ContentPaths;
-	contentTypes: ContentTypes;
-	contentTypesBase: string;
-	hasContentConfig: boolean;
-}) {
-	let contentTypesStr = '';
-	const collectionKeys = Object.keys(contentTypes).sort();
-	for (const collectionKey of collectionKeys) {
-		contentTypesStr += `${collectionKey}: {\n`;
-		const entryKeys = Object.keys(contentTypes[collectionKey]).sort();
-		for (const entryKey of entryKeys) {
-			const entry = contentTypes[collectionKey][entryKey];
-			contentTypesStr += `${entryKey}: ${entry},\n`;
-		}
-		contentTypesStr += `},\n`;
-	}
-
-	let configPathRelativeToCacheDir = normalizePath(
-		path.relative(paths.cacheDir.pathname, paths.config.pathname)
-	);
-	if (!isRelativePath(configPathRelativeToCacheDir))
-		configPathRelativeToCacheDir = './' + configPathRelativeToCacheDir;
-
-	contentTypesBase = contentTypesBase.replace('// @@ENTRY_MAP@@', contentTypesStr);
-	contentTypesBase = contentTypesBase.replace(
-		"'@@CONTENT_CONFIG_TYPE@@'",
-		hasContentConfig ? `typeof import(${JSON.stringify(configPathRelativeToCacheDir)})` : 'never'
-	);
-
-	try {
-		await fs.promises.stat(paths.cacheDir);
-	} catch {
-		await fs.promises.mkdir(paths.cacheDir);
-	}
-
-	await fs.promises.writeFile(new URL(CONTENT_TYPES_FILE, paths.cacheDir), contentTypesBase);
 }
